# Introduce non-formatting log methods