- Add simple benchmarks comparing formatting and non-formatting loggers
- reduce memory allocation by help escape analyzer
- introduce Trace(f), Debug(f), Info(f), Error(f)

Benchmark Results:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkPureMessage/msg=5/context=false/log-4         647           646           -0.15%
BenchmarkPureMessage/msg=5/context=false/logf-4        837           742           -11.35%
BenchmarkPureMessage/msg=5/context=true/log-4          673           682           +1.34%
BenchmarkPureMessage/msg=5/context=true/logf-4         886           720           -18.74%
BenchmarkPureMessage/msg=20/context=false/log-4        644           645           +0.16%
BenchmarkPureMessage/msg=20/context=false/logf-4       848           747           -11.91%
BenchmarkPureMessage/msg=20/context=true/log-4         671           700           +4.32%
BenchmarkPureMessage/msg=20/context=true/logf-4        891           730           -18.07%
BenchmarkPureMessage/msg=100/context=false/log-4       645           642           -0.47%
BenchmarkPureMessage/msg=100/context=false/logf-4      905           804           -11.16%
BenchmarkPureMessage/msg=100/context=true/log-4        673           684           +1.63%
BenchmarkPureMessage/msg=100/context=true/logf-4       954           789           -17.30%
BenchmarkPureMessage/msg=1000/context=false/log-4      646           650           +0.62%
BenchmarkPureMessage/msg=1000/context=false/logf-4     1314          1205          -8.30%
BenchmarkPureMessage/msg=1000/context=true/log-4       673           695           +3.27%
BenchmarkPureMessage/msg=1000/context=true/logf-4      1352          1216          -10.06%

benchmark                                              old allocs     new allocs     delta
BenchmarkPureMessage/msg=5/context=false/log-4         1              1              +0.00%
BenchmarkPureMessage/msg=5/context=false/logf-4        2              0              -100.00%
BenchmarkPureMessage/msg=5/context=true/log-4          1              1              +0.00%
BenchmarkPureMessage/msg=5/context=true/logf-4         3              0              -100.00%
BenchmarkPureMessage/msg=20/context=false/log-4        1              1              +0.00%
BenchmarkPureMessage/msg=20/context=false/logf-4       2              0              -100.00%
BenchmarkPureMessage/msg=20/context=true/log-4         1              1              +0.00%
BenchmarkPureMessage/msg=20/context=true/logf-4        3              0              -100.00%
BenchmarkPureMessage/msg=100/context=false/log-4       1              1              +0.00%
BenchmarkPureMessage/msg=100/context=false/logf-4      2              0              -100.00%
BenchmarkPureMessage/msg=100/context=true/log-4        1              1              +0.00%
BenchmarkPureMessage/msg=100/context=true/logf-4       3              0              -100.00%
BenchmarkPureMessage/msg=1000/context=false/log-4      1              1              +0.00%
BenchmarkPureMessage/msg=1000/context=false/logf-4     2              0              -100.00%
BenchmarkPureMessage/msg=1000/context=true/log-4       1              1              +0.00%
BenchmarkPureMessage/msg=1000/context=true/logf-4      3              0              -100.00%
```